+++ packages/ui/src/components/visual-bell/index.jsx
@@ -1,4 +1,11 @@
-import { UIElement, classnames } from "@elf-framework/sapa";
+import {
+  UIElement,
+  classnames,
+  useEffect,
+  useState,
+  useCallback,
+  potal,
+} from "@elf-framework/sapa";
 
 import { propertyMap } from "../../utils/propertyMap";
 import { makeStyleMap } from "../../utils/styleKeys";
@@ -18,23 +25,56 @@ const cssProperties = makeStyleMap("--elf--visual-bell", {
 
 export class VisualBell extends UIElement {
   template() {
-    const { style = {}, content, direction = "bottom" } = this.props;
+    const { style = {}, content, delay = 0, direction = "bottom" } = this.props;
+    const [localDelay, setLocalDelay] = useState(delay);
+    const [hide, setHide] = useState(false);
+
+    this.state.hideCallback = useCallback(
+      (hideDelay = 0) => {
+        setLocalDelay(hideDelay);
+      },
+      [setLocalDelay]
+    );
 
     const styleObject = {
       class: classnames(
         "elf--visual-bell",
-        `elf--visual-bell-direction-${direction}`
+        `elf--visual-bell-direction-${direction}`,
+        { hide }
       ),
       style: {
         ...propertyMap(style, cssProperties),
+        ...{
+          transition: `opacity ${localDelay}ms ease-in-out`,
+          opacity: hide ? 0 : 1,
+        },
       },
     };
+
+    useEffect(() => {
+      if (localDelay > 0) {
+        // show 상태일 때
+        if (!hide) {
+          this.props.onShow && this.props.onShow();
+        }
+
+        setTimeout(() => {
+          if (!hide) {
+            setHide(true);
+          }
+        }, localDelay);
+      }
+    }, [localDelay, hide]);
 
     return (
       <div
         class="elf--visual-bell"
         {...styleObject}
         onContextMenu={(e) => e.preventDefault()}
+        onTransitionEnd={() => {
+          this.props.onHide && this.props.onHide();
+          this.destroy(true);
+        }}
       >
         <div class="elf--visual-bell-content">
           <div class="elf--visual-bell-text">{content}</div>
@@ -43,4 +83,23 @@ export class VisualBell extends UIElement {
       </div>
     );
   }
+
+  hide(hideDelay = 0) {
+    this.state?.hideCallback(hideDelay);
+  }
+}
+
+export function bell({
+  content = "",
+  delay = 0,
+  direction = "bottom",
+  tools = [],
+  options = {},
+}) {
+  return potal(
+    <VisualBell delay={delay} direction={direction} tools={tools}>
+      {content}
+    </VisualBell>,
+    options
+  );
 }